break up big Main
